开源项目参考/baize-mcp-gateway/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "baize-mcp-gateway"
version = "0.1.0"
description = "白泽 MCP Gateway 服务 - 统一工具调用入口与鉴权"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "pydantic>=2.6",
  "httpx[socks]>=0.27",
  "asyncpg>=0.29",
  "minio>=7.2",
  "uvicorn>=0.30",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "alembic>=1.13",
  "sqlalchemy>=2.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "pre-commit>=3.7",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["baize_mcp_gateway*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
# line-length 已在 [tool.ruff] 顶层配置

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
no_implicit_optional = true
check_untyped_defs = false
disallow_untyped_defs = false
